Move text transformation functions around
